Reorganized external libraries
