+++ UI/flask/Utilities.py
@@ -1,7 +1,1 @@
-from databaseCommunication import *
-from databaseCommunicationUtilities import *
-from Utilities import *
-from SQLFunctions import *
-from config import *
-import sys
