--- UI/flask/databaseCommunicationUtilities.py
@@ -1,8 +1,0 @@
-from databaseCommunication import *
-from databaseCommunicationUtilities import *
-from Utilities import *
-from SQLFunctions import *
-from config import *
-import sys
-
-
