+++ UI/flask/lib/pyVMObject.py
@@ -1,6 +1,5 @@
 """
-
-
+This class represents a VM object.
 
 """
