--- UI/flask/static/style.css.bk
@@ -1,147 +1,0 @@
-* {
-box-sizing: border-box;
-}
-
-ul {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-  overflow: hidden;
-  background-color: #333;
-}
-
-li {
-  float: left;
-    list-style-type: none;
-}
-
-li a {
-  display: block;
-  color: white;
-  /*color: #333333;*/
-  text-align: center;
-  padding: 14px 16px;
-  text-decoration: none;
-}
-
-li a:hover:not(.active) {
-  background-color: #111;
-}
-
-.active {
-  background-color: #4CAF50;
-}
-
-
- 
-*:focus {
-outline: none;
-}
-body {
-  text-align: center;
-font-family: Verdana;
-background-color: #99ccff;
-padding: 50px;
-}
-.login {
-margin: 20px auto;
-width: 300px;
-}
-.login-screen {
-background-color: #FFF;
-padding: 20px;
-}
- 
-.app-title {
-text-align: center;
-color: #e6e6e6;
-    /*color: #777;*/
-}
- 
-.login-form {
-text-align: center;
-}
-.control-group {
-margin-bottom: 10px;
-}
- 
-input {
-text-align: center;
-background-color: #333;
-border: 2px solid transparent;
-border-radius: 3px;
-font-size: 16px;
-font-weight: 200;
-padding: 0px 0;
-width: 250px;
-transition: border .5s;
-}
- 
-input:focus {
-border: 2px solid #3333cc;
-box-shadow: none;
-}
- 
-.btn {
-border: 2px solid transparent;
-background: #99ccff;
-color: #ffffff;
-font-size: 16px;
-line-height: 25px;
-padding: 10px 0;
-text-decoration: none;
-text-shadow: none;
-border-radius: 3px;
-box-shadow: none;
-transition: 0.25s;
-display: block;
-width: 250px;
-margin: 0 auto;
-}
- 
-.btn:hover {
-background-color: #2980B9;
-}
- 
-.login-link {
-font-size: 12px;
-color: #444;
-display: block;
-margin-top: 12px;
-}
-
-
-
-
-/*This is the templates' form in templates.html */
-div.form1
-{
-  display: block;
-  text-align: center;
-}
-
-form.form1
-{
-  display: inline-block;
-  margin-left: 25%;
-  margin-right:25%;
-  width: 50%;
-}
-
-
-/*The bar length thing*/
-.badge {
-  display: inline-block;
-  min-width: 10px;
-  padding: 3px 7px;
-  font-size: 12px;
-  font-weight: bold;
-  line-height: 1;
-  color: #fff;
-  text-align: center;
-  white-space: nowrap;
-  vertical-align: baseline;
-  background-color: #999;
-  border-radius: 10px;
-}
-
